Add popup label when feature is clicked on a Leaflet web map

add_popups creates a popup label containing values of selected variables for a feature selected through an on-click event on a Leaflet web map.

add_popups(in_df, layer_id, label_vars)

Arguments

in_df
A sf data frame object that is displayed on the leaflet web map.

layer_id
The layerId of the feature selected by the user.

label_vars
The variables from in_df selected to be displayed in the popup.

Value

popup_text a html popup label to be displayed on the Leaflet web map.